requirements.txt
flask
flask_cors
pandas
requests
tenacity
sqlalchemy
gunicorn
psycopg2
python-dotenv